% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-release.r
\name{aries.copy.release}
\alias{aries.copy.release}
\title{Create a modified release after removing a set of individuals}
\usage{
aries.copy.release(release.path, new.path, remove.ids)
}
\arguments{
\item{release.path}{Directory path of master release.}

\item{new.path}{Directory path of modified release.}

\item{remove.ids}{ALNQLET identifiers of individuals
who should be omitted from the new release.}
}
\description{
Create a modified release after removing a set of individuals
}
