% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectionp.r
\name{aries.detectionp}
\alias{aries.detectionp}
\title{aries.detectionp}
\usage{
aries.detectionp(selection, start = 1, count = -1)
}
\arguments{
\item{selection}{Output of \code{\link{aries.select}()}.}

\item{start}{Index of the first CpG in the profile (Default: 1).
Indexing is with respect to \code{selection$probe.names}.}

\item{count}{Number of CpG sites to load (Default: -1).
A value of \code{-1} indicates all available.}
}
\value{
A matrix filled with detection p-values (rows=CpG sites, columns=samples).
}
\description{
Load detection p-values for a specified ARIES subset
}
\examples{
\dontrun{
## Load the first 1000 CpG sites for all blood samples collected around age 15
aries <- aries.select(aries.dir, featureset="common", time.point="15up")
aries$detp <- aries.detectionp(aries, start=1, count=1000)
}
}
