% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-sets.r
\name{aries.feature.sets}
\alias{aries.feature.sets}
\title{aries.feature.sets}
\usage{
aries.feature.sets(path)
}
\arguments{
\item{path}{Base directory of ARIES.}
}
\value{
Vector listing available format identifiers.
}
\description{
List of Illumina BeadChip formats available in ARIES
}
