% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.r
\name{aries.info}
\alias{aries.info}
\title{aries.info}
\usage{
aries.info(path)
}
\arguments{
\item{path}{Base directory of ARIES.}
}
\value{
A list containing ARIES information for each Illumina BeadChip format,
run \code{\link{aries.feature.sets}()} to get a list of valid identifiers.
Each of these is in turn a list of items describing the samples
with methylation profiles in that format:
\itemize{
\item samples - Data frame providing sample information, one sample per row.
\item probe.names - Illumina probe identifiers.
\item control.matrix - Matrix providing control probe values (rows=samples, columns=control probes).
\item cell.counts - List of cell count estimates for various references. Each is a matrix (rows=samples, columns=cell types). 
}
}
\description{
Provide comprehensive sample and probe information for all of ARIES
}
\examples{
\dontrun{
aries <- aries.info(aries.dir)
## See CD4+ T cell estimate for the third sample
## with an Illumina EPIC profile.
aries$epic$cell.counts[["blood-gse35069"]][3,"CD4T"]
}
}
