% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.r
\name{aries.samples}
\alias{aries.samples}
\title{aries.samples}
\usage{
aries.samples(path)
}
\arguments{
\item{path}{Base directory of ARIES.}
}
\value{
Data frame listing all samples.
}
\description{
Obtain sample information for all of ARIES.
}
