% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-release.r
\name{aries.save.release}
\alias{aries.save.release}
\title{Saves an ARIES release following QC and normalization}
\usage{
aries.save.release(
  release.path,
  samplesheet,
  qc.objects,
  norm.objects,
  cell.counts,
  qc.summaries,
  norm.summaries,
  qc.report.path,
  norm.report.path,
  betas.path,
  detp.path
)
}
\arguments{
\item{release.path}{Release directory path.}

\item{samplesheet}{Samplesheet for all samples.}

\item{qc.objects}{List of QC objects lists (one for each feature set, i.e. 'common', '450' and 'epic').}

\item{norm.objects}{List of normalised QC object lists corresponding to `qc.objects`.}

\item{cell.counts}{List of cell count estimate matrices (one for each cell count reference).}

\item{qc.summaries}{List of QC report summaries.}

\item{norm.summaries}{List of normalization report summaries.}

\item{qc.report.path}{Directory path of the QC reports.}

\item{norm.report.path}{Directory path of the normalization reports.}

\item{betas.path}{Directory path of the normalized betas GDS files, one for each element in `qc.objects`.}

\item{detp.path}{Directory path of the detection p-value GDS files, one for each element in `qc.objects`.}
}
\description{
Saves an ARIES release following QC and normalization
}
