% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.r
\name{aries.select}
\alias{aries.select}
\title{aries.select}
\usage{
aries.select(
  path,
  featureset,
  time.point = NULL,
  sample.names = NULL,
  replicates = F
)
}
\arguments{
\item{path}{Base directory of ARIES.}

\item{featureset}{Illumina BeadChip format.
#' Run \code{\link{aries.feature.sets}()} to get a list of valid identifiers.  If all selected profiles are from the Illumina 450K (or EPIC) array, then the default is "450" (or "epic").  If there is a mix, then "common" is the default.}

\item{time.point}{Optional. Restrict to profiles for samples collected at specific
time points (Default: NULL).
Run \code{\link{aries.time.points}()} to get a list.}

\item{sample.names}{Optional. Restrict to profiles for a set of samples (Default: NULL).
Run \code{\link{aries.samples}()$Sample_Name} to get a list of sample identifiers.}

\item{replicates}{Include replicate methylation profiles (Default: FALSE).
Ignored if \code{is.null(sample.names)==FALSE}.}
}
\value{
A list containing ARIES information for the ARIES subset:
\itemize{
\item samples - Data frame providing sample information, one sample per row.
\item probe.names - Illumina probe identifiers.
\item control.matrix - Matrix providing control probe values (rows=samples, columns=control probes).
\item cell.counts - List of cell count estimates for various references. Each is a matrix (rows=samples, columns=cell types). 
}
}
\description{
Provide comprehensive sample and probe information for a subset of ARIES
}
\examples{
\dontrun{
## Define subsets of ARIES restricted to samples collected around age 15
## with Illumina 450K and EPIC BeadChip profiles.
aries.15.450 <- aries.select(aries.dir, featureset="450", time.point="15up")
aries.15.epic <- aries.select(aries.dir, featureset="epic", time.point="15up")
}
}
