% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-points.r
\name{aries.time.points}
\alias{aries.time.points}
\title{aries.time.points}
\usage{
aries.time.points(path)
}
\arguments{
\item{path}{Base directory of ARIES.}
}
\value{
A vector of valid time-point identifiers.
}
\description{
List of ARIES time-point identifiers
}
